fix x64 warnings

+++ src/CopyProgress.cpp
@@ -68,7 +68,7 @@ void CopyProgress::DrawTime(int64_t ReadBytes, int64_t WriteBytes, int64_t Total
                             int64_t ReadTime, int64_t WriteTime,
                             int64_t ReadN, int64_t WriteN, int64_t TotalN,
                             int ParallelMode, int64_t FirstWriteTime,
-                            int64_t StartTime, int BufferSize)
+                            int64_t StartTime, size_t BufferSize)
 {
   RedrawWindowIfNeeded();
   double TotalTime     = 0;
@@ -224,7 +224,7 @@ void CopyProgress::ShowProgress(int64_t read, int64_t write, int64_t total,
                                 int64_t readTime, int64_t writeTime,
                                 int64_t readN, int64_t writeN,
                                 int64_t totalN, int parallel,
-                                int64_t FirstWrite, int64_t StartTime, int BufferSize)
+                                int64_t FirstWrite, int64_t StartTime, size_t BufferSize)
 {
   int64_t tm = GetTime();
   if (tm - lastupdate > interval)
